feat: Improved the scalability of RadixSort
Si ens fixem en el fitxer prioritats.json, pujat a l'Estudy, podem veure com
hi ha molts tipus que es sobrepassen del int, es per aixo, que he decidit
passar alguns atributs a long, cosa que ha afcetat al RadixSort, que ha hagut
de ser millorat.

+++ src/main/java/Compare/Comparator.java
@@ -6,5 +6,5 @@ public interface Comparator {
     public boolean compararp1top2 (Post post1, Post post2);
     public boolean compararp2top1 (Post post1, Post post2);
     public boolean compararp2top1IncludeEqual (Post post1, Post post2);
-    public int retornarValor (Post post);
+    public long retornarValor (Post post);
 }
